requirements.txt
pyyaml>=6.0.0
dotenv
jinja2